basic-crud/backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Bcrypt
Flask-Marshmallow
psycopg2-binary
Flask-CORS
marshmallow-sqlalchemy
python-dotenv